Initial refactor

+++ src/EnergyPlus/HVACManager.cc
@@ -2383,7 +2383,7 @@ void ReportAirHeatBalance(EnergyPlusData &state)
         znAirRpt.ExhTotalLoss = 0;
         znAirRpt.ExhSensiLoss = 0;
 
-        for (int FanNum = 1; FanNum <= state.dataFans->fans.size(); ++FanNum) {
+        for (int FanNum = 1; FanNum <= (int)state.dataFans->fans.size(); ++FanNum) {
             auto const *thisFan = state.dataFans->fans(FanNum);
             //  Add reportable vars
             if (thisFan->type == HVAC::FanType::Exhaust) {
